Convert old Timeout's FromMessageItem trait to TryFrom
dbus 0.7 removed FromMessageItem and replaced it with the standard library's
TryFrom. This replicates that behavior for Timeout.

This is a breaking change -- FromMessageItem was a part of Timeout's public
API. However, it will likely have little impact as the dbus connection is an
implementation detail for most users, and is not even available on macos.