Convert NotificationHint's From traits to new MessageItem
Previously, NotificationHints could be converted into MessageItem::DictEntrys.
However, dbus 0.7 removed the MessageItem::DictEntry variant, so dict entries
are now generally represented as tuples of MessageItems where the first one
happens to be a string. Now NotificationHints can be converted to and from
2-tuples of MessageItems.

As a side-effect of the new changes, the util module can be removed.

This is a minor breaking change for non-macos, as some traits implemented on
NotificationHint are removed. Given that macos never even had those traits, it
is unlikely that there will be substantial impact.